Asignando seguridad por roles

+++ msvc-medico/src/main/java/org/proyectosgcs/springcloud/msvc/medico/controllers/EspecialidadController.java
@@ -1,5 +1,7 @@
 package org.proyectosgcs.springcloud.msvc.medico.controllers;
 
+import jakarta.servlet.http.HttpServletRequest;
+import org.proyectosgcs.springcloud.msvc.medico.Auth.JwtAuthorizationHelper;
 import org.proyectosgcs.springcloud.msvc.medico.models.entity.Especialidad;
 import org.proyectosgcs.springcloud.msvc.medico.services.EspecialidadService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,6 +10,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 @RestController
@@ -16,14 +19,34 @@ public class EspecialidadController {
 
     @Autowired
     private EspecialidadService service;
+    @Autowired
+    private JwtAuthorizationHelper jwtAuthorizationHelper;
 
     @GetMapping
-    public List<Especialidad> listar(){
-        return service.listarEspecialidades();
+    public ResponseEntity<?> listar(HttpServletRequest request){
+        if (!jwtAuthorizationHelper.validarRol(request, "ADMIN") &&
+                !jwtAuthorizationHelper.validarRol(request, "MEDICO") &&
+                !jwtAuthorizationHelper.validarRol(request, "PACIENTE")) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(Map.of(
+                    "status", "error",
+                    "message", "Acceso denegado"
+            ));
+        }
+        return ResponseEntity.ok(service.listarEspecialidades());
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<?> detalle(@PathVariable Long id){
+    public ResponseEntity<?> detalle(@PathVariable Long id, HttpServletRequest request){
+
+        if (!jwtAuthorizationHelper.validarRol(request, "ADMIN") &&
+                !jwtAuthorizationHelper.validarRol(request, "MEDICO") &&
+                !jwtAuthorizationHelper.validarRol(request, "PACIENTE")) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(Map.of(
+                    "status", "error",
+                    "message", "Acceso denegado"
+            ));
+        }
+
         Optional<Especialidad> especialidadOptional = service.obtenerEspecialidad(id);
         if (especialidadOptional.isPresent()){
             return ResponseEntity.ok(especialidadOptional.get());
@@ -32,12 +55,26 @@ public class EspecialidadController {
     }
 
     @PostMapping
-    public ResponseEntity<?> crear (@RequestBody Especialidad especialidad){
+    public ResponseEntity<?> crear (@RequestBody Especialidad especialidad, HttpServletRequest request){
+
+        if (!jwtAuthorizationHelper.validarRol(request, "ADMIN")) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(Map.of(
+                    "status", "error",
+                    "message", "Acceso denegado"
+            ));
+        }
+
         return ResponseEntity.status(HttpStatus.CREATED).body(service.registrarEspecialidad(especialidad));
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity <? > editar (@RequestBody Especialidad especialidad, @PathVariable Long id) {
+    public ResponseEntity <? > editar (@RequestBody Especialidad especialidad, @PathVariable Long id, HttpServletRequest request) {
+        if (!jwtAuthorizationHelper.validarRol(request, "ADMIN")) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(Map.of(
+                    "status", "error",
+                    "message", "Acceso denegado"
+            ));
+        }
         Optional<Especialidad> op = service.obtenerEspecialidad(id);
         if (op.isPresent()) {
             Especialidad especialidadDb = op.get();
@@ -48,7 +85,13 @@ public class EspecialidadController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> eliminar(@PathVariable Long id){
+    public ResponseEntity<?> eliminar(@PathVariable Long id, HttpServletRequest request){
+        if (!jwtAuthorizationHelper.validarRol(request, "ADMIN")) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(Map.of(
+                    "status", "error",
+                    "message", "Acceso denegado"
+            ));
+        }
         Optional<Especialidad> op = service.obtenerEspecialidad(id);
         if (op.isPresent()){
             service.eliminarEspecialidad(id);
